+++ StarkBank/StarkBankTests/Event.cs
@@ -1,5 +1,6 @@
 ﻿using Xunit;
 using StarkBank;
+using StarkCore;
 using System;
 using System.Linq;
 using System.Collections.Generic;
@@ -9,7 +10,7 @@ namespace StarkBankTests
 {
     public class EventTest
     {
-        public readonly User user = TestUser.SetDefaultProject();
+        public readonly StarkBank.User user = TestUser.SetDefaultProject();
         public readonly string Content = "{\"event\": {\"log\": {\"transfer\": {\"status\": \"processing\", \"updated\": \"2020-04-03T13:20:33.485644+00:00\", \"fee\": 160, \"name\": \"Lawrence James\", \"accountNumber\": \"10000-0\", \"id\": \"5107489032896512\", \"tags\": [], \"taxId\": \"91.642.017/0001-06\", \"created\": \"2020-04-03T13:20:32.530367+00:00\", \"amount\": 2, \"transactionIds\": [\"6547649079541760\"], \"bankCode\": \"01\", \"branchCode\": \"0001\"}, \"errors\": [], \"type\": \"sending\", \"id\": \"5648419829841920\", \"created\": \"2020-04-03T13:20:33.164373+00:00\"}, \"subscription\": \"transfer\", \"id\": \"6234355449987072\", \"created\": \"2020-04-03T13:20:40.784479+00:00\"}}";
         public readonly string GoodSignature = "MEYCIQCmFCAn2Z+6qEHmf8paI08Ee5ZJ9+KvLWSS3ddp8+RF3AIhALlK7ltfRvMCXhjS7cy8SPlcSlpQtjBxmhN6ClFC0Tv6";
         public readonly string BadSignature = "MEUCIQDOpo1j+V40DNZK2URL2786UQK/8mDXon9ayEd8U0/l7AIgYXtIZJBTs8zCRR3vmted6Ehz/qfw1GRut/eYyvf1yOk=";
@@ -89,7 +90,7 @@ namespace StarkBankTests
         {
             try {
                 Event parsedEvent = Event.Parse(Content, BadSignature);
-            } catch (StarkBank.Error.InvalidSignatureError e) {
+            } catch (StarkCore.Error.InvalidSignatureError e) {
                 TestUtils.Log(e);
                 return;
             }
@@ -103,7 +104,7 @@ namespace StarkBankTests
             {
                 Event parsedEvent = Event.Parse(Content, "something is definitely wrong");
             }
-            catch (StarkBank.Error.InvalidSignatureError e)
+            catch (StarkCore.Error.InvalidSignatureError e)
             {
                 TestUtils.Log(e);
                 return;