+++ StarkBank/StarkBankTests/InvoicePullRequest.cs
@@ -72,6 +72,18 @@ namespace StarkBankTests
             Console.WriteLine("Test: Page Request - OK");
         }
 
+        [Fact]
+        public void TestParseInvoicePullRequestEvent()
+        {
+            string content = "{\"event\": {\"created\": \"2025-07-25T17:36:41.040267+00:00\", \"id\": \"4805265536843776\", \"log\": {\"created\": \"2025-07-25T17:36:39.571648+00:00\", \"description\": \"\", \"errors\": [], \"id\": \"5789040171286528\", \"reason\": \"\", \"request\": {\"attemptType\": \"default\", \"created\": \"2025-07-25T17:36:37.201258+00:00\", \"displayDescription\": \"\", \"due\": \"2025-07-30T07:00:00+00:00\", \"externalId\": \"a15c4821d1c2413a82a4f3cfeee1315e\", \"id\": \"5397390693498880\", \"installmentId\": \"5424937942646784\", \"invoiceId\": \"5118508564217856\", \"status\": \"pending\", \"subscriptionId\": \"5181739848695808\", \"tags\": [], \"updated\": \"2025-07-25T17:36:39.571665+00:00\"}, \"type\": \"pending\"}, \"subscription\": \"invoice-pull-request\", \"workspaceId\": \"6235001133727744\"}}";
+            string validSignature = "MEUCIQCvbPc+mWLLL5nwvOBy/3MVJ3JU9fG/rNmyqmHtaeJA9wIgOR8Tw75MSj7lR9DPqhM62tlq+cFkbw14T4KmDBeC5rM=";
+            Event parsedEvent = Event.Parse(content, validSignature);
+            Assert.NotNull(parsedEvent.ID);
+            Assert.Equal(typeof(InvoicePullRequest.Log), parsedEvent.Log.GetType());
+
+            Console.WriteLine("Test: Parse Invoice Pull Request Event - OK");
+        }
+
         internal static InvoicePullRequest Example(string invoiceId, string subscriptionId)
         {
             return new InvoicePullRequest(