+++ StarkBank/StarkBankTests/InvoicePullSubscription.cs
@@ -119,6 +119,18 @@ namespace StarkBankTests
             Console.WriteLine("Test: Create and Delete Subscription - OK");
         }
 
+        [Fact]
+        public void TestParseInvoicePullSubscriptionEvent()
+        {
+            string content = "{\"event\": {\"created\": \"2025-07-25T17:03:49.207194+00:00\", \"id\": \"5339088045473792\", \"log\": {\"created\": \"2025-07-25T17:03:47.305348+00:00\", \"description\": \"\", \"errors\": [], \"id\": \"4814349822590976\", \"reason\": \"\", \"subscription\": {\"amount\": 1500, \"amountMinLimit\": 0, \"bacenId\": \"RR3990842720250725fwJYIfdOGeF\", \"brcode\": \"00020101021226180014br.gov.bcb.pix5204000053039865802BR5925Stark Sociedade de Credit6009Sao Paulo62070503***80930014br.gov.bcb.pix2571brcode-h.sandbox.starkinfra.com/v2/rec/d2766b29d5184e90853405a9720439a16304686F\", \"created\": \"2025-07-25T17:03:47.280303+00:00\", \"data\": {}, \"displayDescription\": \"fist test - lucas4\", \"due\": \"2025-07-27T17:03:46.709858+00:00\", \"end\": \"2055-06-23T03:00:00+00:00\", \"externalId\": \"3581163bfe96436794a5284d5eb7a5b9\", \"id\": \"4786208928432128\", \"interval\": \"week\", \"name\": \"jaojao\", \"pullMode\": \"manual\", \"pullRetryLimit\": 3, \"referenceCode\": \"ricandalarrapateumanual\", \"start\": \"2055-06-16T03:00:00+00:00\", \"status\": \"created\", \"tags\": [], \"taxId\": \"457.965.518-41\", \"type\": \"qrcode\", \"updated\": \"2025-07-25T17:03:47.305366+00:00\"}, \"type\": \"created\"}, \"subscription\": \"invoice-pull-subscription\", \"workspaceId\": \"6235001133727744\"}}";
+            string validSignature = "MEYCIQDPoI8o1N1qbtq24wY2cvQ4reAxHv/AZs901L6WKU8ylwIhAJ4+ARRrgNARFu/1SbHnuDoHX4EtvkZDCmxTjP9WsT1b";
+            Event parsedEvent = Event.Parse(content, validSignature);
+            Assert.NotNull(parsedEvent.ID);
+            Assert.Equal(typeof(InvoicePullSubscription.Log), parsedEvent.Log.GetType());
+
+            Console.WriteLine("Test: Parse Invoice Pull Subscription Event - OK");
+        }
+
         internal static InvoicePullSubscription Example(string type)
         {
             InvoicePullSubscription example = null;